+++ calculationService/src/main/java/calculation/service/AsyncCalculationService.java
@@ -4,9 +4,11 @@ package calculation.service;
 import calculation.dto.CalculationResponse;
 import org.springframework.web.context.request.async.DeferredResult;
 
+import java.math.BigDecimal;
+
 public interface AsyncCalculationService {
 
-    DeferredResult<CalculationResponse> multipleByTwo(Integer parameter);
+    DeferredResult<CalculationResponse> multipleByTwo(BigDecimal parameter);
 
-    DeferredResult<CalculationResponse> multipleByThree(Integer parameter);
+    DeferredResult<CalculationResponse> multipleByThree(BigDecimal parameter);
 }
